+++ commands/send.js
@@ -1,10 +1,10 @@
-const sending = require('../bot_modules/sendFunction.js');
+const sending = require('../bot_modules/rappelUnban.js');
 
 module.exports = {
-    name : 'send',
-    description : "Méthode pour rappeler l'unban des gens",
-    execute(interaction,client){
-       sending.send(interaction,client);
+    name: 'send',
+    description: "Méthode pour rappeler l'unban des gens",
+    execute(interaction, client) {
+        sending.send(interaction, client);
     }
 }
